extracted ZNO and Entrant structs into separate files
