+++ ZNO.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+namespace StructConsole
+{
+    struct ZNO
+    {
+        public string Subject;
+        public int Points;
+
+        public ZNO(string subject, int points)
+        {
+            Subject = subject;
+            Points = points;
+        }
+    }
+}
